+++ src/vue/components/CommentReply.vue
@@ -9,7 +9,7 @@
         :callback="callback">
       </image-loader>
       <div v-if="!comment.isAttachment(comment.payload.replied_comment_message)" 
-        v-html="escapeHTML(repliedCommentMessage)">
+        v-html="repliedCommentMessage">
       </div>
     </div>
     <image-loader v-if="comment.isAttachment(comment.payload.text)"
@@ -20,7 +20,7 @@
     </image-loader>
     <div class="qcw-comment__content"
       v-if="!comment.isAttachment(comment.payload.text)"
-      v-html="escapeHTML(repliedCommentText)"
+      v-html="repliedCommentText"
     >
     </div>
   </div>
@@ -32,17 +32,6 @@ export default {
   name: 'CommentReply',
   components: {ImageLoader},
   props: ['comment', 'isMe', 'clickHandler', 'repliedCommentSender', 'onClickImage', 'callback', 'repliedCommentMessage', 'repliedCommentText'],
-  methods: {
-    searchAndReplace(str, find, replace) {
-      return str.split(find).join(replace);
-    },
-    escapeHTML(text) {
-      let comment;
-      comment = this.searchAndReplace(text, '<', '&lt;');
-      comment = this.searchAndReplace(comment, '>', '&gt;');
-      return comment;
-    },
-  },
 }
 </script>